refactor(core): change declarative loader filename
